# MasterCompanyApi/DTO/Employee.cs
namespace MasterCompanyApi.DTO
{
    public class Employee
    {
        public string Name { get; set; }
        public string LastName { get; set; }
        public string Document { get; set; }
        public decimal Salary { get; set; }
        public string Gender { get; set; }
        public string Position { get; set; }
        public string StartDate { get; set; }
    }
}