# MasterCompanyApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MasterCompanyApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Employee/increaseSalaries": {
      "get": {
        "tags": [
          "Employee"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employee/genderPercentage": {
      "get": {
        "tags": [
          "Employee"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employee": {
      "get": {
        "tags": [
          "Employee"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employee/salaryRange": {
      "get": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "minSalary",
            "type": "number",
            "format": "double",
            "default": 0
          },
          {
            "in": "query",
            "name": "maxSalary",
            "type": "number",
            "format": "double",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employee/duplicates": {
      "get": {
        "tags": [
          "Employee"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employee/{document}": {
      "delete": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "document",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Employee": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "document": {
          "type": "string"
        },
        "salary": {
          "format": "double",
          "type": "number"
        },
        "gender": {
          "type": "string"
        },
        "position": {
          "type": "string"
        },
        "startDate": {
          "type": "string"
        }
      }
    }
  }
}